% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-advanced-logic.R
\name{add_signature}
\alias{add_signature}
\title{Add Signature}
\usage{
add_signature(name, genes, weights = NULL)
}
\arguments{
\item{name}{Signature name}

\item{genes}{Gene list}

\item{weights}{Gene weights (optional)}
}
\value{
Signature definition
}
\description{
Add Signature
}
