% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-core.R
\name{add_stat_annotation}
\alias{add_stat_annotation}
\title{Add Statistical Annotation}
\usage{
add_stat_annotation(p, test_result, x = NULL, y = NULL)
}
\arguments{
\item{p}{ggplot object}

\item{test_result}{Statistical test result}

\item{x}{X position for annotation}

\item{y}{Y position for annotation}
}
\value{
ggplot object with annotation
}
\description{
Add statistical test results to plots.
}
\keyword{internal}
