% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-ga-general.R
\name{analysis-ga-general}
\alias{analysis-ga-general}
\title{General Analysis Functions for Shiny Modules}
\description{
This file contains general analysis functions extracted from UCSCXenaShiny
for use in ZinaSuite Shiny applications. These functions are specifically
designed for the General Analysis module.
}
