% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-advanced-logic.R
\name{analyze_cross_gene}
\alias{analyze_cross_gene}
\title{Analyze Cross Gene}
\usage{
analyze_cross_gene(target_gene, correlate_genes, cancer = NULL)
}
\arguments{
\item{target_gene}{Target gene}

\item{correlate_genes}{Genes to correlate with}

\item{cancer}{Cancer type}
}
\value{
Cross-gene analysis result
}
\description{
Analyze Cross Gene
}
