% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-advanced-logic.R
\name{analyze_cross_pathway}
\alias{analyze_cross_pathway}
\title{Analyze Cross Pathway}
\usage{
analyze_cross_pathway(pathway1, pathway2, cancer = NULL)
}
\arguments{
\item{pathway1}{First pathway genes}

\item{pathway2}{Second pathway genes}

\item{cancer}{Cancer type}
}
\value{
Cross-pathway analysis result
}
\description{
Analyze Cross Pathway
}
