% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-advanced-logic.R
\name{analyze_dimension_distribution}
\alias{analyze_dimension_distribution}
\title{Analyze Dimension Distribution}
\usage{
analyze_dimension_distribution(gene, cancer = NULL)
}
\arguments{
\item{gene}{Gene symbol}

\item{cancer}{Cancer type}
}
\value{
Distribution analysis
}
\description{
Analyze Dimension Distribution
}
