% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-advanced-logic.R
\name{analyze_dimension_reduction}
\alias{analyze_dimension_reduction}
\title{Analyze Dimension Reduction}
\usage{
analyze_dimension_reduction(
  genes,
  method = "pca",
  cancer = NULL,
  n_components = 2
)
}
\arguments{
\item{genes}{Gene list}

\item{method}{Method: "pca", "tsne", "umap"}

\item{cancer}{Cancer type}

\item{n_components}{Number of components}
}
\value{
Dimension reduction result
}
\description{
Analyze Dimension Reduction
}
