% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-advanced-logic.R
\name{analyze_drug_omic}
\alias{analyze_drug_omic}
\title{Analyze Drug Omic Pair}
\usage{
analyze_drug_omic(drug, omic_type, feature)
}
\arguments{
\item{drug}{Drug name}

\item{omic_type}{Omic data type}

\item{feature}{Feature name}
}
\value{
Drug-omic pair analysis
}
\description{
Analyze Drug Omic Pair
}
