% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-advanced-logic.R
\name{analyze_drug_response}
\alias{analyze_drug_response}
\title{Analyze Drug Response}
\usage{
analyze_drug_response(genes, drug, cancer = NULL)
}
\arguments{
\item{genes}{Gene list}

\item{drug}{Drug name}

\item{cancer}{Cancer type}
}
\value{
Drug response analysis
}
\description{
Analyze Drug Response
}
