% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-advanced-logic.R
\name{analyze_drug_sensitivity}
\alias{analyze_drug_sensitivity}
\title{Analyze Drug Sensitivity}
\usage{
analyze_drug_sensitivity(gene, drug, cancer = NULL)
}
\arguments{
\item{gene}{Gene symbol}

\item{drug}{Drug name}

\item{cancer}{Cancer type}
}
\value{
Drug sensitivity analysis
}
\description{
Analyze Drug Sensitivity
}
