% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-advanced-logic.R
\name{analyze_feature_across}
\alias{analyze_feature_across}
\title{Analyze Feature Across Types}
\usage{
analyze_feature_across(feature, types = NULL)
}
\arguments{
\item{feature}{Feature name}

\item{types}{Cancer types to compare}
}
\value{
Cross-cancer analysis
}
\description{
Analyze Feature Across Types
}
