% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-advanced-logic.R
\name{analyze_feature_signature}
\alias{analyze_feature_signature}
\title{Analyze Feature Signature}
\usage{
analyze_feature_signature(features, signature_name, cancer = NULL)
}
\arguments{
\item{features}{Feature list}

\item{signature_name}{Signature name}

\item{cancer}{Cancer type}
}
\value{
Signature analysis
}
\description{
Analyze Feature Signature
}
