% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-advanced-logic.R
\name{analyze_group_comparison}
\alias{analyze_group_comparison}
\title{Analyze Group Comparison}
\usage{
analyze_group_comparison(gene, group_var, cancer = NULL)
}
\arguments{
\item{gene}{Gene symbol}

\item{group_var}{Grouping variable}

\item{cancer}{Cancer type}
}
\value{
Group comparison result
}
\description{
Analyze Group Comparison
}
