% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-advanced-logic.R
\name{analyze_pathway}
\alias{analyze_pathway}
\title{Analyze Pathway}
\usage{
analyze_pathway(pathway_name, gene_list, cancer = NULL)
}
\arguments{
\item{pathway_name}{Pathway name}

\item{gene_list}{List of genes in pathway}

\item{cancer}{Cancer type}
}
\value{
Pathway analysis result
}
\description{
Analyze Pathway
}
