% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-grouping.R
\name{apply_filter}
\alias{apply_filter}
\title{Apply Filter to Metadata}
\usage{
apply_filter(data, var, values, operator)
}
\arguments{
\item{data}{Data frame}

\item{var}{Variable name}

\item{values}{Filter values}

\item{operator}{Filter operator}
}
\value{
Filtered data frame
}
\description{
Apply Filter to Metadata
}
\keyword{internal}
