% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-logic.R
\name{build_analysis_params}
\alias{build_analysis_params}
\title{Build Analysis Parameters}
\usage{
build_analysis_params(analysis_type, gene1, gene2, cancer = NULL)
}
\arguments{
\item{analysis_type}{Type of analysis}

\item{gene1}{First gene}

\item{gene2}{Second gene}

\item{cancer}{Cancer type}
}
\value{
List of parameters
}
\description{
Build Analysis Parameters
}
