% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-logic.R
\name{build_plot_params}
\alias{build_plot_params}
\title{Build Plot Parameters}
\usage{
build_plot_params(
  plot_type,
  title = NULL,
  x_label = NULL,
  y_label = NULL,
  color = "default"
)
}
\arguments{
\item{plot_type}{Plot type}

\item{title}{Plot title}

\item{x_label}{X-axis label}

\item{y_label}{Y-axis label}

\item{color}{Color scheme}
}
\value{
List of parameters
}
\description{
Build Plot Parameters
}
