% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-logic.R
\name{build_query_params}
\alias{build_query_params}
\title{Build Query Parameters}
\usage{
build_query_params(gene, data_type, source)
}
\arguments{
\item{gene}{Gene symbol}

\item{data_type}{Data type (mRNA, mutation, cnv, etc.)}

\item{source}{Data source (tcga, pcawg, ccle)}
}
\value{
List of query parameters
}
\description{
Build Query Parameters
}
