% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cache.R
\name{cached_query}
\alias{cached_query}
\title{Cache-aware Data Query}
\usage{
cached_query(query_func, cache_key, use_cache = TRUE, ttl = 3600)
}
\arguments{
\item{query_func}{Function to execute if cache miss}

\item{cache_key}{Cache key}

\item{use_cache}{Whether to use cache}

\item{ttl}{Cache TTL in seconds}
}
\value{
Query result
}
\description{
Query data with automatic caching
}
