% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-logic.R
\name{calculate_progress_pct}
\alias{calculate_progress_pct}
\title{Calculate Progress Percentage}
\usage{
calculate_progress_pct(tracker)
}
\arguments{
\item{tracker}{Progress tracker}
}
\value{
Percentage (0-100)
}
\description{
Calculate Progress Percentage
}
