% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccle_absolute}
\alias{ccle_absolute}
\title{CCLE ABSOLUTE Data}
\format{
A data frame with ABSOLUTE results
}
\source{
UCSC Xena Public Hub
}
\description{
ABSOLUTE purity and ploidy estimates for CCLE cell lines.
}
\examples{
data("ccle_absolute")
head(ccle_absolute)
}
