% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccle_info}
\alias{ccle_info}
\title{CCLE Sample Information}
\format{
A data frame with CCLE cell line metadata
}
\source{
UCSC Xena Public Hub
}
\description{
Phenotype data for CCLE cell lines.
}
\examples{
data("ccle_info")
head(ccle_info)
}
