% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccle_info_fine}
\alias{ccle_info_fine}
\title{CCLE Sample Information (Fine-grained)}
\format{
A data frame with cleaned CCLE variables
}
\source{
UCSC Xena Public Hub
}
\description{
Cleaned CCLE phenotype data for grouping.
}
\examples{
data("ccle_info_fine")
head(ccle_info_fine)
}
