% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ccle.R
\name{CCLEData}
\alias{CCLEData}
\title{CCLEData R6 Class}
\description{
R6 class for accessing CCLE (Cancer Cell Line Encyclopedia) data.
}
\examples{
\dontrun{
# Create CCLE data instance
ccle <- CCLEData$new()

# Query gene expression
tp53_expr <- ccle$get_gene_expression("TP53")

# Query mutation status
tp53_mut <- ccle$get_mutation_status("TP53")
}
}
\section{Super class}{
\code{\link[ZinaSuite:DataSource]{ZinaSuite::DataSource}} -> \code{CCLEData}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CCLEData-new}{\code{CCLEData$new()}}
\item \href{#method-CCLEData-get_gene_expression}{\code{CCLEData$get_gene_expression()}}
\item \href{#method-CCLEData-get_mutation_status}{\code{CCLEData$get_mutation_status()}}
\item \href{#method-CCLEData-get_cnv}{\code{CCLEData$get_cnv()}}
\item \href{#method-CCLEData-get_protein_expression}{\code{CCLEData$get_protein_expression()}}
\item \href{#method-CCLEData-query}{\code{CCLEData$query()}}
\item \href{#method-CCLEData-clone}{\code{CCLEData$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ZinaSuite" data-topic="DataSource" data-id="clear_cache"><a href='../../ZinaSuite/html/DataSource.html#method-DataSource-clear_cache'><code>ZinaSuite::DataSource$clear_cache()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ZinaSuite" data-topic="DataSource" data-id="get_info"><a href='../../ZinaSuite/html/DataSource.html#method-DataSource-get_info'><code>ZinaSuite::DataSource$get_info()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ZinaSuite" data-topic="DataSource" data-id="query_batch"><a href='../../ZinaSuite/html/DataSource.html#method-DataSource-query_batch'><code>ZinaSuite::DataSource$query_batch()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CCLEData-new"></a>}}
\if{latex}{\out{\hypertarget{method-CCLEData-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new CCLEData instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CCLEData$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CCLEData-get_gene_expression"></a>}}
\if{latex}{\out{\hypertarget{method-CCLEData-get_gene_expression}{}}}
\subsection{Method \code{get_gene_expression()}}{
Get gene expression data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CCLEData$get_gene_expression(gene)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gene}}{Gene symbol}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Named numeric vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CCLEData-get_mutation_status"></a>}}
\if{latex}{\out{\hypertarget{method-CCLEData-get_mutation_status}{}}}
\subsection{Method \code{get_mutation_status()}}{
Get mutation status
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CCLEData$get_mutation_status(gene)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gene}}{Gene symbol}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Named numeric vector (0/1)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CCLEData-get_cnv"></a>}}
\if{latex}{\out{\hypertarget{method-CCLEData-get_cnv}{}}}
\subsection{Method \code{get_cnv()}}{
Get CNV data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CCLEData$get_cnv(gene)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gene}}{Gene symbol}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Named numeric vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CCLEData-get_protein_expression"></a>}}
\if{latex}{\out{\hypertarget{method-CCLEData-get_protein_expression}{}}}
\subsection{Method \code{get_protein_expression()}}{
Get protein expression data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CCLEData$get_protein_expression(protein)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{protein}}{Protein identifier}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Named numeric vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CCLEData-query"></a>}}
\if{latex}{\out{\hypertarget{method-CCLEData-query}{}}}
\subsection{Method \code{query()}}{
Query data by type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CCLEData$query(identifier, data_type = "expression")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{identifier}}{Gene or protein symbol}

\item{\code{data_type}}{Type of data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Named numeric vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CCLEData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CCLEData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CCLEData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
