% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-check.R
\name{check_shiny_deps}
\alias{check_shiny_deps}
\title{Check Shiny Dependencies}
\usage{
check_shiny_deps()
}
\value{
TRUE if all packages are installed
}
\description{
Check if all Shiny-related dependencies are installed.
}
\examples{
\dontrun{
check_shiny_deps()
}
}
\keyword{internal}
