% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-check.R
\name{check_suggested_pkgs}
\alias{check_suggested_pkgs}
\title{Check Multiple Suggested Packages}
\usage{
check_suggested_pkgs(pkgs, reason = NULL)
}
\arguments{
\item{pkgs}{Vector of package names}

\item{reason}{Optional reason why the packages are needed}
}
\value{
TRUE if all packages are installed
}
\description{
Check if multiple suggested packages are installed.
}
\examples{
\dontrun{
check_suggested_pkgs(c("survival", "survminer"))
}
}
\keyword{internal}
