% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-logic.R
\name{create_batch_job}
\alias{create_batch_job}
\title{Create Batch Job}
\usage{
create_batch_job(job_type, params, priority = 1)
}
\arguments{
\item{job_type}{Job type}

\item{params}{Job parameters}

\item{priority}{Job priority}
}
\value{
Batch job object
}
\description{
Create Batch Job
}
