% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-logic.R
\name{create_notification}
\alias{create_notification}
\title{Create Notification Message}
\usage{
create_notification(type, title, message)
}
\arguments{
\item{type}{Notification type}

\item{title}{Notification title}

\item{message}{Notification message}
}
\value{
Notification object
}
\description{
Create Notification Message
}
