% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-logic.R
\name{create_query_dataframe}
\alias{create_query_dataframe}
\title{Create Query Result DataFrame}
\usage{
create_query_dataframe(query_result)
}
\arguments{
\item{query_result}{Query result}
}
\value{
Data frame
}
\description{
Create Query Result DataFrame
}
