% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-logic.R
\name{create_status}
\alias{create_status}
\title{Create Status Message}
\usage{
create_status(status, message, details = NULL)
}
\arguments{
\item{status}{Status type}

\item{message}{Message text}

\item{details}{Additional details}
}
\value{
Status object
}
\description{
Create Status Message
}
