% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-advanced-logic.R
\name{custom_metadata}
\alias{custom_metadata}
\title{Custom Metadata}
\usage{
custom_metadata(metadata, sample_col = "sample")
}
\arguments{
\item{metadata}{Data frame with metadata}

\item{sample_col}{Column containing sample IDs}
}
\value{
Processed metadata
}
\description{
Custom Metadata
}
