% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-DataSource.R
\name{DataSource}
\alias{DataSource}
\title{DataSource R6 Class}
\description{
Abstract base class for data sources. Provides common interface for querying
molecular data from various sources (TCGA, PCAWG, CCLE, etc.).
}
\examples{
\dontrun{
# This is an abstract class - use concrete implementations
# See XenaData class for actual usage
}

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataSource-new}{\code{DataSource$new()}}
\item \href{#method-DataSource-query}{\code{DataSource$query()}}
\item \href{#method-DataSource-query_batch}{\code{DataSource$query_batch()}}
\item \href{#method-DataSource-get_info}{\code{DataSource$get_info()}}
\item \href{#method-DataSource-clear_cache}{\code{DataSource$clear_cache()}}
\item \href{#method-DataSource-clone}{\code{DataSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataSource-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new DataSource instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSource$new(name, host = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Data source name}

\item{\code{host}}{Host URL or identifier}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSource-query"></a>}}
\if{latex}{\out{\hypertarget{method-DataSource-query}{}}}
\subsection{Method \code{query()}}{
Query data for a single identifier
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSource$query(identifier, data_type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{identifier}}{Gene symbol, transcript ID, or other molecular identifier}

\item{\code{data_type}}{Type of data (e.g., "mRNA", "protein", "mutation")}

\item{\code{...}}{Additional parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data vector or NULL if not found
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSource-query_batch"></a>}}
\if{latex}{\out{\hypertarget{method-DataSource-query_batch}{}}}
\subsection{Method \code{query_batch()}}{
Query data for multiple identifiers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSource$query_batch(
  identifiers,
  data_type,
  async = FALSE,
  n_workers = 4,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{identifiers}}{Vector of molecular identifiers}

\item{\code{data_type}}{Type of data}

\item{\code{async}}{Whether to use async/parallel processing}

\item{\code{n_workers}}{Number of workers for async processing}

\item{\code{...}}{Additional parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Named list of query results
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSource-get_info"></a>}}
\if{latex}{\out{\hypertarget{method-DataSource-get_info}{}}}
\subsection{Method \code{get_info()}}{
Get data source information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSource$get_info()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List with data source details
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSource-clear_cache"></a>}}
\if{latex}{\out{\hypertarget{method-DataSource-clear_cache}{}}}
\subsection{Method \code{clear_cache()}}{
Clear the cache
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSource$clear_cache()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Invisible self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
