% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-id-utils.R
\name{determine_tcga_cancer}
\alias{determine_tcga_cancer}
\title{Determine TCGA Cancer Type}
\usage{
determine_tcga_cancer(id)
}
\arguments{
\item{id}{Sample ID or vector of sample IDs}
}
\value{
Cancer type abbreviation (e.g., "BRCA", "LUAD")
}
\description{
Extract cancer type from TCGA sample ID.
Uses the tissue source site (TSS) code to map to cancer type.
}
\examples{
\dontrun{
determine_tcga_cancer("TCGA-19-1787-01")  # Returns: "GBM"
}
}
