% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{.ZinaSuiteEnv}
\alias{.ZinaSuiteEnv}
\title{ZinaSuite Package Environment}
\format{
An object of class \code{environment} of length 2.
}
\usage{
.ZinaSuiteEnv
}
\description{
ZinaSuite Package Environment
}
\keyword{internal}
