% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-logic.R
\name{execute_batch_job}
\alias{execute_batch_job}
\title{Execute Batch Job}
\usage{
execute_batch_job(job)
}
\arguments{
\item{job}{Batch job}
}
\value{
Updated job with result
}
\description{
Execute Batch Job
}
