% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-logic.R
\name{execute_correlation_analysis}
\alias{execute_correlation_analysis}
\title{Execute Correlation Analysis}
\usage{
execute_correlation_analysis(gene1_data, gene2_data)
}
\arguments{
\item{gene1_data}{Expression data for gene 1}

\item{gene2_data}{Expression data for gene 2}
}
\value{
Analysis result
}
\description{
Execute Correlation Analysis
}
