% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-logic.R
\name{execute_data_query}
\alias{execute_data_query}
\title{Execute Data Query}
\usage{
execute_data_query(gene, data_type, source)
}
\arguments{
\item{gene}{Gene symbol}

\item{data_type}{Data type}

\item{source}{Data source}
}
\value{
Query result or error
}
\description{
Execute Data Query
}
