% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-id-utils.R
\name{extract_barcode}
\alias{extract_barcode}
\title{Extract Barcode from Sample ID}
\usage{
extract_barcode(id)
}
\arguments{
\item{id}{Sample ID or vector of sample IDs}
}
\value{
Barcode(s) (first 12 characters)
}
\description{
Extract the first 12 characters (TCGA-XX-XXXX) from TCGA sample ID.
This is used for matching with clinical data.
}
\examples{
\dontrun{
extract_barcode("TCGA-19-1787-01A-01R-1564-08")
# Returns: "TCGA-19-1787"
}
}
