% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-advanced-logic.R
\name{filter_samples}
\alias{filter_samples}
\title{Filter Samples}
\usage{
filter_samples(samples, filters = list())
}
\arguments{
\item{samples}{Sample IDs}

\item{filters}{List of filters}
}
\value{
Filtered samples
}
\description{
Filter Samples
}
