% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-id-utils.R
\name{filter_tcga_normal}
\alias{filter_tcga_normal}
\title{Filter TCGA Normal Samples}
\usage{
filter_tcga_normal(data)
}
\arguments{
\item{data}{Named vector or data frame with TCGA sample IDs}
}
\value{
Filtered data with only normal samples
}
\description{
Filter TCGA data to keep only normal samples (sample type 10-19).
}
\examples{
\dontrun{
data <- query_gene_expression("TP53")
normal_data <- filter_tcga_normal(data)
}
}
