% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-id-utils.R
\name{filter_tcga_tumor}
\alias{filter_tcga_tumor}
\title{Filter TCGA Tumor Samples}
\usage{
filter_tcga_tumor(data)
}
\arguments{
\item{data}{Named vector or data frame with TCGA sample IDs}
}
\value{
Filtered data with only tumor samples
}
\description{
Filter TCGA data to keep only tumor samples (sample type 01-09).
}
\examples{
\dontrun{
data <- query_gene_expression("TP53")
tumor_data <- filter_tcga_tumor(data)
}
}
