% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-logic.R
\name{format_batch_results}
\alias{format_batch_results}
\title{Format Batch Results}
\usage{
format_batch_results(jobs)
}
\arguments{
\item{jobs}{List of batch jobs}
}
\value{
Summary data frame
}
\description{
Format Batch Results
}
