% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_bytes}
\alias{format_bytes}
\title{Format bytes to human-readable string}
\usage{
format_bytes(bytes)
}
\arguments{
\item{bytes}{Number of bytes}
}
\value{
Formatted string
}
\description{
Format bytes to human-readable string
}
\keyword{internal}
