% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-logic.R
\name{format_progress}
\alias{format_progress}
\title{Format Progress Message}
\usage{
format_progress(tracker)
}
\arguments{
\item{tracker}{Progress tracker}
}
\value{
Formatted string
}
\description{
Format Progress Message
}
