% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-logic.R
\name{format_query_results}
\alias{format_query_results}
\title{Format Query Results}
\usage{
format_query_results(query_result)
}
\arguments{
\item{query_result}{Query result from execute_data_query}
}
\value{
Formatted summary
}
\description{
Format Query Results
}
