% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-logic.R
\name{format_status}
\alias{format_status}
\title{Format Status for Display}
\usage{
format_status(status)
}
\arguments{
\item{status}{Status object}
}
\value{
Formatted string
}
\description{
Format Status for Display
}
