% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-ga-general.R
\name{ga_p_survplot}
\alias{ga_p_survplot}
\title{Helper function for survival plotting}
\usage{
ga_p_survplot(df, palette, ...)
}
\description{
Helper function for survival plotting
}
\keyword{internal}
