% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-ga-general.R
\name{ga_sur_plot}
\alias{ga_sur_plot}
\title{Helper function for survival plotting with cutoff}
\usage{
ga_sur_plot(df, cutoff_mode, cutpoint, palette, ...)
}
\description{
Helper function for survival plotting with cutoff
}
\keyword{internal}
