% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-logic.R
\name{generate_correlation_plot}
\alias{generate_correlation_plot}
\title{Generate Correlation Plot}
\usage{
generate_correlation_plot(x, y, title = "Correlation")
}
\arguments{
\item{x}{First variable}

\item{y}{Second variable}

\item{title}{Plot title}
}
\value{
ggplot object
}
\description{
Generate Correlation Plot
}
