% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-logic.R
\name{generate_distribution_plot}
\alias{generate_distribution_plot}
\title{Generate Distribution Plot}
\usage{
generate_distribution_plot(
  data,
  title = "Distribution",
  plot_type = "histogram"
)
}
\arguments{
\item{data}{Numeric vector}

\item{title}{Plot title}

\item{plot_type}{Plot type}
}
\value{
ggplot object
}
\description{
Generate Distribution Plot
}
