% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-generation.R
\name{generate_report_template}
\alias{generate_report_template}
\title{Generate Report Template}
\usage{
generate_report_template(data, title, author, format)
}
\arguments{
\item{data}{Analysis data}

\item{title}{Report title}

\item{author}{Report author}

\item{format}{Output format}
}
\value{
Rmd template string
}
\description{
Generate Report Template
}
\keyword{internal}
