% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-app.R
\name{get_app_dir}
\alias{get_app_dir}
\title{Get Shiny App Directory}
\usage{
get_app_dir()
}
\value{
Character string with the app directory path
}
\description{
Get the file path to the Shiny application directory.
}
\examples{
\dontrun{
app_dir <- get_app_dir()
list.files(app_dir)
}
}
