% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_async_engine}
\alias{get_async_engine}
\title{Get or create global async compute engine}
\usage{
get_async_engine(n_workers = parallel::detectCores() - 1)
}
\arguments{
\item{n_workers}{Number of workers}
}
\value{
AsyncCompute instance
}
\description{
Get or create global async compute engine
}
\keyword{internal}
