% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-pharmacogenomics.R
\name{get_available_drugs}
\alias{get_available_drugs}
\title{Get Available Drugs}
\usage{
get_available_drugs(source = c("ccle", "gdsc"))
}
\arguments{
\item{source}{Data source}
}
\value{
Character vector of drug names
}
\description{
Get list of available drugs for pharmacogenomic analysis.
}
\examples{
\dontrun{
drugs <- get_available_drugs()
head(drugs)
}
}
