% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data_source}
\alias{get_data_source}
\title{Get data source instance}
\usage{
get_data_source(source = c("tcga", "pcawg", "ccle"))
}
\arguments{
\item{source}{Data source name: "tcga", "pcawg", "ccle"}
}
\value{
DataSource instance
}
\description{
Get data source instance
}
\examples{
\dontrun{
tcga_source <- get_data_source("tcga")
}
}
