% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-advanced-logic.R
\name{get_gene_description}
\alias{get_gene_description}
\title{Get Gene Description}
\usage{
get_gene_description(gene)
}
\arguments{
\item{gene}{Gene symbol}
}
\value{
Gene description
}
\description{
Get Gene Description
}
\keyword{internal}
