% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-advanced-logic.R
\name{get_gene_related_cancers}
\alias{get_gene_related_cancers}
\title{Get Gene Related Cancers}
\usage{
get_gene_related_cancers(gene)
}
\arguments{
\item{gene}{Gene symbol}
}
\value{
Related cancers
}
\description{
Get Gene Related Cancers
}
\keyword{internal}
