% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-grouping.R
\name{get_grouping_variables}
\alias{get_grouping_variables}
\title{Get Available Grouping Variables}
\usage{
get_grouping_variables(database = c("tcga", "pcawg", "ccle"))
}
\arguments{
\item{database}{Data source: "tcga", "pcawg", or "ccle"}
}
\value{
Character vector of available grouping variables
}
\description{
List all available grouping variables for a database.
}
\examples{
\dontrun{
# Get TCGA grouping variables
vars <- get_grouping_variables("tcga")
head(vars)

# Get PCAWG variables
pcawg_vars <- get_grouping_variables("pcawg")
}
}
