% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-anatomy.R
\name{get_organ_info}
\alias{get_organ_info}
\title{Get Organ Information for Cancer Types}
\usage{
get_organ_info(cancers = NULL)
}
\arguments{
\item{cancers}{Vector of cancer types. If NULL, returns all.}
}
\value{
Data frame with Cancer, Organ, and System columns
}
\description{
Retrieve anatomical and system information for TCGA cancer types.
}
\examples{
\donttest{
# Get organ info for all cancers
organ_info <- get_organ_info()

# Get specific cancer info
brca_info <- get_organ_info("BRCA")
}
}
