% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-core.R
\name{get_palette}
\alias{get_palette}
\title{Get Color Palette}
\usage{
get_palette(n, palette = c("default", "cancer", "gradient", "diverging"))
}
\arguments{
\item{n}{Number of colors}

\item{palette}{Palette name: "default", "cancer", "gradient"}
}
\value{
Vector of colors
}
\description{
Get a color palette for visualizations.
}
\examples{
# Get 5 colors from default palette
colors <- get_palette(5)

# Get gradient palette
grad_colors <- get_palette(10, palette = "gradient")
}
