% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinasuite-data.R
\name{get_pancan_cn_value}
\alias{get_pancan_cn_value}
\title{Query Copy Number Variation (CNV)}
\usage{
get_pancan_cn_value(identifier, use_thresholded = TRUE, host = "tcgaHub")
}
\arguments{
\item{identifier}{Gene symbol}

\item{use_thresholded}{Use thresholded data (default: TRUE)}

\item{host}{Xena hub host (default: "toilHub")}
}
\value{
Named numeric vector of GISTIC2 scores
}
\description{
Query Copy Number Variation (CNV)
}
