% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinasuite-data.R
\name{get_pancan_mirna_value}
\alias{get_pancan_mirna_value}
\title{Query miRNA Expression}
\usage{
get_pancan_mirna_value(identifier, host = "toilHub")
}
\arguments{
\item{identifier}{miRNA ID (e.g., "hsa-miR-21-5p")}

\item{host}{Xena hub host (default: "toilHub")}
}
\value{
Named numeric vector of miRNA expression
}
\description{
Query miRNA Expression
}
