% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinasuite-data.R
\name{get_pancan_mutation_status}
\alias{get_pancan_mutation_status}
\title{Query Mutation Status}
\usage{
get_pancan_mutation_status(identifier, host = "pancanAtlasHub")
}
\arguments{
\item{identifier}{Gene symbol}

\item{host}{Xena hub host (default: "toilHub")}
}
\value{
Named numeric vector (0 = wild-type, 1 = mutated)
}
\description{
Query Mutation Status
}
