% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinasuite-data.R
\name{get_pancan_protein_value}
\alias{get_pancan_protein_value}
\title{Query Protein Expression (RPPA)}
\usage{
get_pancan_protein_value(identifier, host = "toilHub")
}
\arguments{
\item{identifier}{Protein identifier}

\item{host}{Xena hub host (default: "toilHub")}
}
\value{
Named numeric vector of protein expression values
}
\description{
Query Protein Expression (RPPA)
}
