% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinasuite-data.R
\name{get_pancan_transcript_value}
\alias{get_pancan_transcript_value}
\title{Query Transcript Expression}
\usage{
get_pancan_transcript_value(
  identifier,
  norm = c("tpm", "fpkm"),
  host = "toilHub"
)
}
\arguments{
\item{identifier}{Transcript ID or gene symbol}

\item{norm}{Normalization method: "tpm" or "fpkm"}

\item{host}{Xena hub host (default: "toilHub")}
}
\value{
Named numeric vector of transcript expression
}
\description{
Query Transcript Expression
}
