% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cache.R, R/zzz.R
\name{get_zina_cache}
\alias{get_zina_cache}
\title{Get Global Cache Instance}
\usage{
get_zina_cache()

get_zina_cache()
}
\value{
ZinaCache object

ZinaCache instance
}
\description{
Get or create the global ZinaCache instance
}
