% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-ccle-drug.R
\name{gm_mean}
\alias{gm_mean}
\title{Geometric Mean}
\usage{
gm_mean(x, na.rm = TRUE, zero.propagate = FALSE)
}
\arguments{
\item{x}{Numeric vector}

\item{na.rm}{Remove NAs (default: TRUE)}

\item{zero.propagate}{Propagate zeros (default: FALSE)}
}
\value{
Geometric mean
}
\description{
Calculate geometric mean of a vector.
}
\keyword{internal}
