% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-advanced-logic.R
\name{help_identifier}
\alias{help_identifier}
\title{Help Identifier}
\usage{
help_identifier(id)
}
\arguments{
\item{id}{Identifier to look up}
}
\value{
Help information
}
\description{
Help Identifier
}
