% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_gene}
\alias{is_valid_gene}
\title{Validate gene symbol}
\usage{
is_valid_gene(gene)
}
\arguments{
\item{gene}{Gene symbol to validate}
}
\value{
TRUE if valid, FALSE otherwise
}
\description{
Validate gene symbol
}
\keyword{internal}
