% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_samples}
\alias{is_valid_samples}
\title{Validate sample IDs}
\usage{
is_valid_samples(sample_ids)
}
\arguments{
\item{sample_ids}{Character vector of sample IDs}
}
\value{
TRUE if valid, FALSE otherwise
}
\description{
Validate sample IDs
}
\keyword{internal}
