% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_progress_callback}
\alias{make_progress_callback}
\title{Create a progress callback function}
\usage{
make_progress_callback(total, message = "Processing")
}
\arguments{
\item{total}{Total number of items}

\item{message}{Progress message}
}
\value{
Callback function
}
\description{
Create a progress callback function
}
\keyword{internal}
