% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-id-utils.R
\name{match_samples}
\alias{match_samples}
\title{Match Samples Between Different Data Sources}
\usage{
match_samples(ids1, ids2, source1 = "auto", source2 = "auto")
}
\arguments{
\item{ids1}{First set of sample IDs}

\item{ids2}{Second set of sample IDs}

\item{source1}{Source of first set ("tcga", "ccle", "pcawg", "auto")}

\item{source2}{Source of second set ("tcga", "ccle", "pcawg", "auto")}
}
\value{
List with matched IDs and indices
}
\description{
Match samples between two datasets with potentially different ID formats.
Returns the intersection of standardized sample IDs.
}
\examples{
\dontrun{
ids1 <- c("TCGA-19-1787-01A", "TCGA-B2-5641-11A")
ids2 <- c("TCGA-19-1787-01", "TCGA-B2-5641-11")
match <- match_samples(ids1, ids2)
}
}
