% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-grouping.R
\name{merge_categorical_groups}
\alias{merge_categorical_groups}
\title{Merge Categorical Groups}
\usage{
merge_categorical_groups(values, merge_by)
}
\arguments{
\item{values}{Character values}

\item{merge_by}{Merge definition}
}
\value{
Character vector of merged group labels
}
\description{
Merge Categorical Groups
}
\keyword{internal}
