% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-grouping.R
\name{merge_groups}
\alias{merge_groups}
\title{Merge Groups}
\usage{
merge_groups(data, group, merge_by, merge_quantile)
}
\arguments{
\item{data}{Data frame}

\item{group}{Group variable name}

\item{merge_by}{Merge definition}

\item{merge_quantile}{Whether to use quantiles}
}
\value{
Data frame with merged groups
}
\description{
Merge Groups
}
\keyword{internal}
