% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-grouping.R
\name{merge_numeric_groups}
\alias{merge_numeric_groups}
\title{Merge Numeric Groups}
\usage{
merge_numeric_groups(values, cancers, merge_by, use_quantile)
}
\arguments{
\item{values}{Numeric values}

\item{cancers}{Cancer types}

\item{merge_by}{Merge definition}

\item{use_quantile}{Whether to use quantiles}
}
\value{
Character vector of merged group labels
}
\description{
Merge Numeric Groups
}
\keyword{internal}
