% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null-coalescing}
\alias{null-coalescing}
\alias{\%||\%}
\alias{null_coalesce}
\title{Null-default operator}
\usage{
lhs \%||\% rhs

lhs \%||\% rhs
}
\arguments{
\item{lhs}{Left-hand side value}

\item{rhs}{Right-hand side value}
}
\value{
lhs if not NULL, otherwise rhs

lhs if not NULL, otherwise rhs
}
\description{
Returns the left-hand side if not NULL, otherwise returns the right-hand side.
}
\keyword{internal}
