% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-advanced-logic.R
\name{pancan_search}
\alias{pancan_search}
\title{Pan-Cancer Search}
\usage{
pancan_search(query)
}
\arguments{
\item{query}{Search query (gene, pathway, or cancer)}
}
\value{
Search results
}
\description{
Pan-Cancer Search
}
