% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcawg_info}
\alias{pcawg_info}
\title{PCAWG Sample Information}
\format{
A data frame with PCAWG sample metadata
}
\source{
UCSC Xena PCAWG Hub
}
\description{
Phenotype data for PCAWG samples.
}
\examples{
data("pcawg_info")
head(pcawg_info)
}
