% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcawg_info_fine}
\alias{pcawg_info_fine}
\title{PCAWG Sample Information (Fine-grained)}
\format{
A data frame with cleaned PCAWG variables
}
\source{
UCSC Xena PCAWG Hub
}
\description{
Cleaned PCAWG phenotype data for grouping.
}
\examples{
data("pcawg_info_fine")
head(pcawg_info_fine)
}
