% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcawg_purity}
\alias{pcawg_purity}
\title{PCAWG Purity Data}
\format{
A data frame with purity scores
}
\source{
UCSC Xena PCAWG Hub
}
\description{
Tumor purity estimates for PCAWG samples.
}
\examples{
data("pcawg_purity")
head(pcawg_purity)
}
