% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pcawg.R
\name{PCAWGData}
\alias{PCAWGData}
\title{PCAWGData R6 Class}
\description{
R6 class for accessing PCAWG (Pan-Cancer Analysis of Whole Genomes) data.
}
\examples{
\dontrun{
# Create PCAWG data instance
pcawg <- PCAWGData$new()

# Query gene expression
tp53_expr <- pcawg$get_gene_expression("TP53")

# Query mutation status
tp53_mut <- pcawg$get_mutation_status("TP53")
}
}
\section{Super class}{
\code{\link[ZinaSuite:DataSource]{ZinaSuite::DataSource}} -> \code{PCAWGData}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PCAWGData-new}{\code{PCAWGData$new()}}
\item \href{#method-PCAWGData-get_gene_expression}{\code{PCAWGData$get_gene_expression()}}
\item \href{#method-PCAWGData-get_mutation_status}{\code{PCAWGData$get_mutation_status()}}
\item \href{#method-PCAWGData-get_cnv}{\code{PCAWGData$get_cnv()}}
\item \href{#method-PCAWGData-query}{\code{PCAWGData$query()}}
\item \href{#method-PCAWGData-clone}{\code{PCAWGData$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ZinaSuite" data-topic="DataSource" data-id="clear_cache"><a href='../../ZinaSuite/html/DataSource.html#method-DataSource-clear_cache'><code>ZinaSuite::DataSource$clear_cache()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ZinaSuite" data-topic="DataSource" data-id="get_info"><a href='../../ZinaSuite/html/DataSource.html#method-DataSource-get_info'><code>ZinaSuite::DataSource$get_info()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ZinaSuite" data-topic="DataSource" data-id="query_batch"><a href='../../ZinaSuite/html/DataSource.html#method-DataSource-query_batch'><code>ZinaSuite::DataSource$query_batch()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PCAWGData-new"></a>}}
\if{latex}{\out{\hypertarget{method-PCAWGData-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PCAWGData instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PCAWGData$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PCAWGData-get_gene_expression"></a>}}
\if{latex}{\out{\hypertarget{method-PCAWGData-get_gene_expression}{}}}
\subsection{Method \code{get_gene_expression()}}{
Get gene expression data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PCAWGData$get_gene_expression(gene)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gene}}{Gene symbol}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Named numeric vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PCAWGData-get_mutation_status"></a>}}
\if{latex}{\out{\hypertarget{method-PCAWGData-get_mutation_status}{}}}
\subsection{Method \code{get_mutation_status()}}{
Get mutation status
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PCAWGData$get_mutation_status(gene)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gene}}{Gene symbol}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Named numeric vector (0/1)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PCAWGData-get_cnv"></a>}}
\if{latex}{\out{\hypertarget{method-PCAWGData-get_cnv}{}}}
\subsection{Method \code{get_cnv()}}{
Get CNV data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PCAWGData$get_cnv(gene)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gene}}{Gene symbol}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Named numeric vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PCAWGData-query"></a>}}
\if{latex}{\out{\hypertarget{method-PCAWGData-query}{}}}
\subsection{Method \code{query()}}{
Query data by type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PCAWGData$query(identifier, data_type = "expression")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{identifier}}{Gene symbol}

\item{\code{data_type}}{Type of data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Named numeric vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PCAWGData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PCAWGData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PCAWGData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
