% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-anatomy.R
\name{plot_anatomy_barplot}
\alias{plot_anatomy_barplot}
\title{Plot Anatomy Barplot}
\usage{
plot_anatomy_barplot(data, gene, color_palette, text_size)
}
\arguments{
\item{data}{Plot data frame}

\item{gene}{Gene name}

\item{color_palette}{Color palette}

\item{text_size}{Text size}
}
\value{
ggplot object
}
\description{
Plot Anatomy Barplot
}
\keyword{internal}
