% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-immune.R
\name{plot_immune_barplot}
\alias{plot_immune_barplot}
\title{Plot Immune Correlation Barplot}
\usage{
plot_immune_barplot(data, gene)
}
\arguments{
\item{data}{Correlation results data frame}

\item{gene}{Gene name for title}
}
\value{
ggplot object
}
\description{
Plot Immune Correlation Barplot
}
\keyword{internal}
