% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-immune.R
\name{plot_immune_dotplot}
\alias{plot_immune_dotplot}
\title{Plot Immune Correlation Dotplot}
\usage{
plot_immune_dotplot(data, gene)
}
\arguments{
\item{data}{Correlation results data frame}

\item{gene}{Gene name for title}
}
\value{
ggplot object
}
\description{
Plot Immune Correlation Dotplot
}
\keyword{internal}
