% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-validation.R
\name{print.DataQualityReport}
\alias{print.DataQualityReport}
\title{Print Data Quality Report}
\usage{
\method{print}{DataQualityReport}(x, ...)
}
\arguments{
\item{x}{DataQualityReport object}

\item{...}{Additional arguments}
}
\description{
Print Data Quality Report
}
