% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ccle.R
\name{query_ccle_cnv}
\alias{query_ccle_cnv}
\title{Query CCLE Copy Number Variation}
\usage{
query_ccle_cnv(gene, host = "publicHub")
}
\arguments{
\item{gene}{Gene symbol}

\item{host}{CCLE hub host (default: "publicHub")}
}
\value{
Named numeric vector of CNV values
}
\description{
Query CCLE Copy Number Variation
}
\examples{
\dontrun{
myc_cnv <- query_ccle_cnv("MYC")
summary(myc_cnv)
}
}
