% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ccle.R
\name{query_ccle_expression}
\alias{query_ccle_expression}
\title{Query CCLE Gene Expression}
\usage{
query_ccle_expression(gene, host = "publicHub")
}
\arguments{
\item{gene}{Gene symbol}

\item{host}{CCLE hub host (default: "publicHub")}
}
\value{
Named numeric vector of gene expression values
}
\description{
Query CCLE Gene Expression
}
\examples{
\dontrun{
tp53_expr <- query_ccle_expression("TP53")
head(tp53_expr)
}
}
