% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ccle.R
\name{query_ccle_mutation}
\alias{query_ccle_mutation}
\title{Query CCLE Mutation Status}
\usage{
query_ccle_mutation(gene, host = "publicHub")
}
\arguments{
\item{gene}{Gene symbol}

\item{host}{CCLE hub host (default: "publicHub")}
}
\value{
Named numeric vector (0 = wild-type, 1 = mutated)
}
\description{
Query CCLE Mutation Status
}
\examples{
\dontrun{
tp53_mut <- query_ccle_mutation("TP53")
table(tp53_mut)
}
}
