% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ccle.R
\name{query_ccle_protein}
\alias{query_ccle_protein}
\title{Query CCLE Protein Expression (RPPA)}
\usage{
query_ccle_protein(protein, host = "publicHub")
}
\arguments{
\item{protein}{Protein identifier}

\item{host}{CCLE hub host (default: "publicHub")}
}
\value{
Named numeric vector of protein expression values
}
\description{
Query CCLE Protein Expression (RPPA)
}
\examples{
\dontrun{
tp53_prot <- query_ccle_protein("TP53")
head(tp53_prot)
}
}
