% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-load.R
\name{query_immune_infiltration}
\alias{query_immune_infiltration}
\title{Query Immune Infiltration Data}
\usage{
query_immune_infiltration(cell_type = NULL, source = "tcga")
}
\arguments{
\item{cell_type}{Immune cell type to query}

\item{source}{Data source (default: "tcga")}
}
\value{
Named vector of infiltration scores
}
\description{
Query immune cell infiltration data.
}
\examples{
\dontrun{
# Query CD8 T cell infiltration
cd8 <- query_immune_infiltration("CD8_T_cell", source = "tcga")
summary(cd8)
}
}
