% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-load.R
\name{query_msi}
\alias{query_msi}
\title{Query MSI Data}
\usage{
query_msi(source = "tcga")
}
\arguments{
\item{source}{Data source (default: "tcga")}
}
\value{
Named vector of MSI scores
}
\description{
Query microsatellite instability (MSI) data.
}
\examples{
\dontrun{
# Query MSI
msi <- query_msi(source = "tcga")
table(msi)
}
}
