% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pcawg.R
\name{query_pcawg_cnv}
\alias{query_pcawg_cnv}
\title{Query PCAWG Copy Number Variation}
\usage{
query_pcawg_cnv(gene, host = "pcawgHub")
}
\arguments{
\item{gene}{Gene symbol}

\item{host}{PCAWG hub host (default: "pcawgHub")}
}
\value{
Named numeric vector of CNV values
}
\description{
Query PCAWG Copy Number Variation
}
\examples{
\dontrun{
myc_cnv <- query_pcawg_cnv("MYC")
summary(myc_cnv)
}
}
