% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pcawg.R
\name{query_pcawg_expression}
\alias{query_pcawg_expression}
\title{Query PCAWG Gene Expression}
\usage{
query_pcawg_expression(gene, host = "pcawgHub")
}
\arguments{
\item{gene}{Gene symbol}

\item{host}{PCAWG hub host (default: "pcawgHub")}
}
\value{
Named numeric vector of gene expression values
}
\description{
Query PCAWG Gene Expression
}
\examples{
\dontrun{
tp53_expr <- query_pcawg_expression("TP53")
head(tp53_expr)
}
}
