% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pcawg.R
\name{query_pcawg_mutation}
\alias{query_pcawg_mutation}
\title{Query PCAWG Mutation Status}
\usage{
query_pcawg_mutation(gene, host = "pcawgHub")
}
\arguments{
\item{gene}{Gene symbol}

\item{host}{PCAWG hub host (default: "pcawgHub")}
}
\value{
Named numeric vector (0 = wild-type, 1 = mutated)
}
\description{
Query PCAWG Mutation Status
}
\examples{
\dontrun{
tp53_mut <- query_pcawg_mutation("TP53")
table(tp53_mut)
}
}
