% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-load.R
\name{query_til}
\alias{query_til}
\title{Query TIL Data}
\usage{
query_til(cell_type = NULL, source = "tcga")
}
\arguments{
\item{cell_type}{Cell type to query (default: NULL for all)}

\item{source}{Data source (default: "tcga")}
}
\value{
Data frame of TIL fractions
}
\description{
Query tumor-infiltrating lymphocyte (TIL) data.
}
\examples{
\dontrun{
# Query all TIL data
til <- query_til(source = "tcga")
head(til)
}
}
