% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{retry}
\alias{retry}
\title{Retry a function call with exponential backoff}
\usage{
retry(expr, max_tries = 3, initial_delay = 1)
}
\arguments{
\item{expr}{Expression to evaluate}

\item{max_tries}{Maximum number of attempts}

\item{initial_delay}{Initial delay in seconds}
}
\value{
Result of expression evaluation
}
\description{
Retry a function call with exponential backoff
}
\keyword{internal}
