% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-generation.R
\name{run_pipeline}
\alias{run_pipeline}
\title{Run Analysis Pipeline}
\usage{
run_pipeline(pipeline, ...)
}
\arguments{
\item{pipeline}{AnalysisPipeline object}

\item{...}{Arguments passed to first step}
}
\value{
Pipeline results
}
\description{
Run Analysis Pipeline
}
