% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-tcga-pipeline.R
\name{run_tcga_comp_m2o}
\alias{run_tcga_comp_m2o}
\title{Run TCGA Comparison Analysis - Many-to-One}
\usage{
run_tcga_comp_m2o(molecules_text, cancer, data_type)
}
\arguments{
\item{molecules_text}{Text with multiple molecules}

\item{cancer}{Cancer type code}

\item{data_type}{Data type}
}
\value{
List containing plot, data, and stats
}
\description{
Run TCGA Comparison Analysis - Many-to-One
}
