% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-tcga-pipeline.R
\name{run_tcga_comp_o2m}
\alias{run_tcga_comp_o2m}
\title{Run TCGA Comparison Analysis - One-to-Many}
\usage{
run_tcga_comp_o2m(molecule, data_type)
}
\arguments{
\item{molecule}{Molecule identifier}

\item{data_type}{Data type}
}
\value{
List containing plot, data, and stats
}
\description{
Run TCGA Comparison Analysis - One-to-Many
}
