% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-tcga-pipeline.R
\name{run_tcga_comp_o2o}
\alias{run_tcga_comp_o2o}
\title{Run TCGA Comparison Analysis - One-to-One}
\usage{
run_tcga_comp_o2o(molecule, cancer, data_type)
}
\arguments{
\item{molecule}{Molecule identifier}

\item{cancer}{Cancer type code}

\item{data_type}{Data type}
}
\value{
List containing plot, data, and stats
}
\description{
Run TCGA Comparison Analysis - One-to-One
}
