% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-tcga-pipeline.R
\name{run_tcga_cor_m2o}
\alias{run_tcga_cor_m2o}
\title{Run TCGA Correlation Analysis - Many-to-One}
\usage{
run_tcga_cor_m2o(molecules_text, cancer, data_type, cor_method)
}
\arguments{
\item{molecules_text}{Text with multiple molecules (one per line)}

\item{cancer}{Cancer type code}

\item{data_type}{Data type}

\item{cor_method}{Correlation method}
}
\value{
List containing plot, data, and stats
}
\description{
Run TCGA Correlation Analysis - Many-to-One
}
