% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-tcga-pipeline.R
\name{run_tcga_cor_o2m}
\alias{run_tcga_cor_o2m}
\title{Run TCGA Correlation Analysis - One-to-Many}
\usage{
run_tcga_cor_o2m(molecule1, molecule2, data_type1, data_type2, cor_method)
}
\arguments{
\item{molecule1}{First molecule identifier}

\item{molecule2}{Second molecule identifier}

\item{data_type1}{Data type for first molecule}

\item{data_type2}{Data type for second molecule}

\item{cor_method}{Correlation method}
}
\value{
List containing plot, data, and stats
}
\description{
Run TCGA Correlation Analysis - One-to-Many
}
