% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-tcga-pipeline.R
\name{run_tcga_cross_o2m}
\alias{run_tcga_cross_o2m}
\title{Run TCGA Cross-Omics Analysis - One-to-Many}
\usage{
run_tcga_cross_o2m(molecule1, molecule2, cancer, data_type1, data_type2)
}
\arguments{
\item{molecule1}{First molecule identifier}

\item{molecule2}{Second molecule identifier}

\item{cancer}{Cancer type code}

\item{data_type1}{Data type for first molecule}

\item{data_type2}{Data type for second molecule}
}
\value{
List containing plot, data, and stats
}
\description{
Run TCGA Cross-Omics Analysis - One-to-Many
}
