% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-tcga-pipeline.R
\name{run_tcga_sur_m2o}
\alias{run_tcga_sur_m2o}
\title{Run TCGA Survival Analysis - Many-to-One}
\usage{
run_tcga_sur_m2o(
  molecules_text,
  cancer,
  data_type,
  surv_measure = "OS",
  cutoff_mode = "Auto"
)
}
\arguments{
\item{molecules_text}{Text with multiple molecules}

\item{cancer}{Cancer type code}

\item{data_type}{Data type}

\item{surv_measure}{Survival measure}

\item{cutoff_mode}{Cutoff mode}
}
\value{
List containing plot, data, and stats
}
\description{
Run TCGA Survival Analysis - Many-to-One
}
