% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-tcga-pipeline.R
\name{run_tcga_sur_o2m}
\alias{run_tcga_sur_o2m}
\title{Run TCGA Survival Analysis - One-to-Many}
\usage{
run_tcga_sur_o2m(
  molecule,
  data_type,
  surv_measure = "OS",
  cutoff_mode = "Auto"
)
}
\arguments{
\item{molecule}{Molecule identifier}

\item{data_type}{Data type}

\item{surv_measure}{Survival measure}

\item{cutoff_mode}{Cutoff mode}
}
\value{
List containing plot, data, and stats
}
\description{
Run TCGA Survival Analysis - One-to-Many
}
