% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-core.R
\name{save_plot}
\alias{save_plot}
\title{Save Plot}
\usage{
save_plot(plot, filename, width = 8, height = 6, dpi = 300, ...)
}
\arguments{
\item{plot}{ggplot object}

\item{filename}{Output filename}

\item{width}{Width in inches}

\item{height}{Height in inches}

\item{dpi}{Resolution}

\item{...}{Additional arguments to ggsave}
}
\description{
Save a ggplot with consistent settings.
}
\examples{
\dontrun{
p <- ggplot(mtcars, aes(x = mpg, y = wt)) + geom_point()
save_plot(p, "myplot.pdf", width = 8, height = 6)
}
}
