% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-advanced-logic.R
\name{shiny-advanced-logic}
\alias{shiny-advanced-logic}
\title{Advanced Shiny Business Logic}
\description{
Advanced business logic functions for Shiny modules covering all UCSCXenaShiny features.
}
