% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-modules.R
\name{shiny-module-server}
\alias{shiny-module-server}
\alias{mod_home_server}
\alias{mod_data_query_server}
\alias{mod_analysis_server}
\alias{mod_visualization_server}
\alias{mod_pancan_server}
\alias{mod_mutation_server}
\alias{mod_dimension_server}
\alias{mod_immune_server}
\alias{mod_pharmacogenomics_server}
\alias{mod_batch_server}
\alias{mod_about_server}
\alias{mod_quick_tcga_server}
\title{Shiny Module Server Functions}
\usage{
mod_home_server(id)

mod_data_query_server(id, app_state)

mod_analysis_server(id, app_state)

mod_visualization_server(id, app_state)

mod_pancan_server(id, app_state)

mod_mutation_server(id, app_state)

mod_dimension_server(id, app_state)

mod_immune_server(id, app_state)

mod_pharmacogenomics_server(id, app_state)

mod_batch_server(id, app_state, async_compute)

mod_about_server(id)

mod_quick_tcga_server(id, app_state, async_compute)
}
\description{
Server functions for all Shiny modules in ZinaSuite.
}
\keyword{internal}
