% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-modules.R
\name{shiny-module-ui}
\alias{shiny-module-ui}
\alias{mod_home_ui}
\alias{mod_data_query_ui}
\alias{mod_analysis_ui}
\alias{mod_visualization_ui}
\alias{mod_pancan_ui}
\alias{mod_mutation_ui}
\alias{mod_dimension_ui}
\alias{mod_immune_ui}
\alias{mod_pharmacogenomics_ui}
\alias{mod_batch_ui}
\alias{mod_about_ui}
\title{Shiny Module UI Functions}
\usage{
mod_home_ui(id)

mod_data_query_ui(id)

mod_analysis_ui(id)

mod_visualization_ui(id)

mod_pancan_ui(id)

mod_mutation_ui(id)

mod_dimension_ui(id)

mod_immune_ui(id)

mod_pharmacogenomics_ui(id)

mod_batch_ui(id)

mod_about_ui(id)
}
\description{
UI functions for all Shiny modules in ZinaSuite.
These are re-exported from the inst/shinyapp/modules/ directory.
}
\keyword{internal}
