% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA.organ}
\alias{TCGA.organ}
\title{TCGA Organ Data}
\format{
A data frame with organ mappings
}
\source{
UCSC Xena
}
\description{
Anatomical location mapping for TCGA cancer types.
}
\examples{
data("TCGA.organ")
head(TCGA.organ)
}
