% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tcga_barcode_to_cancer}
\alias{tcga_barcode_to_cancer}
\title{Extract cancer type from TCGA barcode}
\usage{
tcga_barcode_to_cancer(barcode)
}
\arguments{
\item{barcode}{TCGA barcode}
}
\value{
Cancer type abbreviation
}
\description{
Extract cancer type from TCGA barcode
}
\keyword{internal}
