% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tcga_barcode_to_type}
\alias{tcga_barcode_to_type}
\title{Convert TCGA barcode to sample type}
\usage{
tcga_barcode_to_type(barcode)
}
\arguments{
\item{barcode}{TCGA barcode}
}
\value{
Sample type: "tumor" or "normal"
}
\description{
Convert TCGA barcode to sample type
}
\keyword{internal}
