% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_clinical}
\alias{tcga_clinical}
\title{TCGA Clinical Data}
\format{
A data frame with TCGA clinical information
}
\source{
UCSC Xena Toil Hub
}
\description{
Clinical phenotype data for TCGA samples.
}
\examples{
data("tcga_clinical")
head(tcga_clinical)
}
