% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_clinical_fine}
\alias{tcga_clinical_fine}
\title{TCGA Clinical Data (Fine-grained)}
\format{
A data frame with cleaned clinical variables
}
\source{
UCSC Xena Toil Hub
}
\description{
Cleaned TCGA clinical data optimized for sample grouping.
}
\examples{
data("tcga_clinical_fine")
head(tcga_clinical_fine)
}
