% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_genome_instability}
\alias{tcga_genome_instability}
\title{TCGA Genome Instability Data}
\format{
A data frame with instability scores
}
\source{
GDC Pan-Cancer Atlas
}
\description{
Genome instability metrics for TCGA samples.
}
\examples{
data("tcga_genome_instability")
head(tcga_genome_instability)
}
