% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_gtex}
\alias{tcga_gtex}
\title{TCGA GTEx Sample Information}
\format{
A data frame with sample metadata
}
\source{
UCSC Xena Toil Hub
}
\description{
Sample information for TCGA and GTEx samples.
}
\examples{
data("tcga_gtex")
head(tcga_gtex)
}
