% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_MSI}
\alias{tcga_MSI}
\title{TCGA MSI Data}
\format{
A data frame with MSI scores
}
\source{
GDC Pan-Cancer Atlas
}
\description{
Microsatellite instability (MSI) scores for TCGA samples.
}
\examples{
data("tcga_MSI")
head(tcga_MSI)
}
