% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_purity}
\alias{tcga_purity}
\title{TCGA Purity Data}
\format{
A data frame with purity scores
}
\source{
Aran et al., Nat Commun 2015
}
\description{
Tumor purity estimates for TCGA samples.
}
\examples{
data("tcga_purity")
head(tcga_purity)
}
