% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_stemness}
\alias{tcga_stemness}
\title{TCGA Stemness Data}
\format{
A data frame with stemness scores
}
\source{
Malta et al., Cell 2018
}
\description{
Tumor stemness scores for TCGA samples.
}
\examples{
data("tcga_stemness")
head(tcga_stemness)
}
