% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_subtypes}
\alias{tcga_subtypes}
\title{TCGA Subtype Data}
\format{
A data frame with subtype assignments
}
\source{
UCSC Xena
}
\description{
Molecular subtype classifications for TCGA samples.
}
\examples{
data("tcga_subtypes")
head(tcga_subtypes)
}
