% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_surv}
\alias{tcga_surv}
\title{TCGA Survival Data}
\format{
A data frame with OS and PFS information
}
\source{
UCSC Xena Toil Hub
}
\description{
Overall survival data for TCGA samples.
}
\examples{
data("tcga_surv")
head(tcga_surv)
}
