% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_TIL}
\alias{tcga_TIL}
\title{TCGA TIL Data}
\format{
A data frame with TIL fractions
}
\source{
Thorsson et al., Immunity 2018
}
\description{
Tumor-infiltrating lymphocyte (TIL) fractions for TCGA samples.
}
\examples{
data("tcga_TIL")
head(tcga_TIL)
}
