% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_tmb}
\alias{tcga_tmb}
\title{TCGA TMB Data}
\format{
A data frame with TMB scores
}
\source{
GDC Pan-Cancer Atlas
}
\description{
Tumor mutation burden (TMB) for TCGA samples.
}
\examples{
data("tcga_tmb")
head(tcga_tmb)
}
