% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-load.R
\name{try_load_remote_data}
\alias{try_load_remote_data}
\title{Try to Load Remote Data}
\usage{
try_load_remote_data(data_name, force_download = FALSE)
}
\arguments{
\item{data_name}{Name of the dataset}

\item{force_download}{Force re-download}
}
\value{
Data frame or NULL if not available
}
\description{
Attempt to download data from remote sources.
}
\keyword{internal}
