% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinasuite-data.R
\name{try_query_value}
\alias{try_query_value}
\title{Try Query Value with Retry Logic}
\usage{
try_query_value(host, dataset, identifier, max_try = 5L, use_probeMap = TRUE)
}
\arguments{
\item{host}{Xena hub host}

\item{dataset}{Dataset name}

\item{identifier}{Gene/protein identifier}

\item{max_try}{Maximum retry attempts}
}
\value{
Data matrix or NULL if failed
}
\description{
Try Query Value with Retry Logic
}
\keyword{internal}
