% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-logic.R
\name{update_progress}
\alias{update_progress}
\title{Update Progress}
\usage{
update_progress(tracker, step, message = NULL)
}
\arguments{
\item{tracker}{Progress tracker}

\item{step}{Step number}

\item{message}{Step message}
}
\value{
Updated tracker
}
\description{
Update Progress
}
