% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-logic.R
\name{validate_analysis_input}
\alias{validate_analysis_input}
\title{Validate Analysis Input}
\usage{
validate_analysis_input(analysis_type, gene1, gene2)
}
\arguments{
\item{analysis_type}{Analysis type}

\item{gene1}{Gene 1}

\item{gene2}{Gene 2}
}
\value{
List with valid and message
}
\description{
Validate Analysis Input
}
