% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-validation.R
\name{validate_cancer_type}
\alias{validate_cancer_type}
\title{Validate Cancer Type}
\usage{
validate_cancer_type(cancer, valid_cancers = NULL)
}
\arguments{
\item{cancer}{Character vector of cancer types}

\item{valid_cancers}{Optional vector of valid cancer types}
}
\value{
TRUE if valid, throws error otherwise
}
\description{
Validate Cancer Type
}
\examples{
validate_cancer_type("BRCA")
validate_cancer_type(c("BRCA", "LUAD", "LUSC"))
}
