% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-core.R
\name{validate_plot_data}
\alias{validate_plot_data}
\title{Validate Plot Data}
\usage{
validate_plot_data(data, required_cols)
}
\arguments{
\item{data}{Data frame to validate}

\item{required_cols}{Required column names}
}
\description{
Internal function to validate data before plotting.
}
\keyword{internal}
