% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-logic.R
\name{validate_plot_input}
\alias{validate_plot_input}
\title{Validate Plot Input}
\usage{
validate_plot_input(data, plot_type)
}
\arguments{
\item{data}{Data to plot}

\item{plot_type}{Plot type}
}
\value{
List with valid and message
}
\description{
Validate Plot Input
}
