% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-logic.R
\name{validate_query_input}
\alias{validate_query_input}
\title{Validate Query Input}
\usage{
validate_query_input(gene, data_type, source)
}
\arguments{
\item{gene}{Gene symbol}

\item{data_type}{Data type}

\item{source}{Data source}
}
\value{
List with valid (TRUE/FALSE) and message
}
\description{
Validate Query Input
}
