% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-validation.R
\name{validate_sample_ids}
\alias{validate_sample_ids}
\title{Validate Sample IDs}
\usage{
validate_sample_ids(sample_ids, pattern = NULL)
}
\arguments{
\item{sample_ids}{Character vector of sample IDs}

\item{pattern}{Regex pattern for validation}
}
\value{
TRUE if valid, throws error otherwise
}
\description{
Validate Sample IDs
}
\examples{
validate_sample_ids(c("TCGA-01-1234-01", "TCGA-02-5678-01"))
}
