% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-advanced-logic.R
\name{validate_uploaded_data}
\alias{validate_uploaded_data}
\title{Validate Uploaded Data}
\usage{
validate_uploaded_data(data, expected_format = "gene_list")
}
\arguments{
\item{data}{Uploaded data}

\item{expected_format}{Expected format}
}
\value{
Validation result
}
\description{
Validate Uploaded Data
}
