% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-survival.R
\name{vis_kaplan_meier}
\alias{vis_kaplan_meier}
\title{Kaplan-Meier Survival Curve}
\usage{
vis_kaplan_meier(km_result, title = NULL, conf.int = TRUE, risk.table = FALSE)
}
\arguments{
\item{km_result}{KM analysis result from analyze_kaplan_meier}

\item{title}{Plot title}

\item{conf.int}{Whether to show confidence intervals}

\item{risk.table}{Whether to show risk table}
}
\value{
ggplot object or ggsurvplot object
}
\description{
Create a Kaplan-Meier survival curve plot.
}
\keyword{internal}
