% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-survival.R
\name{vis_kaplan_meier_basic}
\alias{vis_kaplan_meier_basic}
\title{Basic Kaplan-Meier plot using ggplot2}
\usage{
vis_kaplan_meier_basic(fit, title = NULL, conf.int = TRUE, pvalue = NULL)
}
\description{
Basic Kaplan-Meier plot using ggplot2
}
\keyword{internal}
