% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cache.R
\name{ZinaCache}
\alias{ZinaCache}
\title{Data Cache Management}
\description{
Advanced data caching system for ZinaSuite to improve performance
and reduce redundant data queries.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{cache_dir}}{Directory for disk cache}

\item{\code{max_memory_items}}{Maximum items in memory cache}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ZinaCache-new}{\code{ZinaCache$new()}}
\item \href{#method-ZinaCache-set}{\code{ZinaCache$set()}}
\item \href{#method-ZinaCache-get}{\code{ZinaCache$get()}}
\item \href{#method-ZinaCache-has}{\code{ZinaCache$has()}}
\item \href{#method-ZinaCache-remove}{\code{ZinaCache$remove()}}
\item \href{#method-ZinaCache-clear}{\code{ZinaCache$clear()}}
\item \href{#method-ZinaCache-stats}{\code{ZinaCache$stats()}}
\item \href{#method-ZinaCache-keys}{\code{ZinaCache$keys()}}
\item \href{#method-ZinaCache-clone}{\code{ZinaCache$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZinaCache-new"></a>}}
\if{latex}{\out{\hypertarget{method-ZinaCache-new}{}}}
\subsection{Method \code{new()}}{
Initialize cache manager
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZinaCache$new(cache_dir = NULL, max_memory_items = 100)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cache_dir}}{Cache directory path}

\item{\code{max_memory_items}}{Maximum memory cache size}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZinaCache-set"></a>}}
\if{latex}{\out{\hypertarget{method-ZinaCache-set}{}}}
\subsection{Method \code{set()}}{
Store data in cache
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZinaCache$set(key, data, use_disk = FALSE, ttl = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Cache key}

\item{\code{data}}{Data to cache}

\item{\code{use_disk}}{Whether to cache to disk}

\item{\code{ttl}}{Time to live in seconds (NULL = no expiration)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZinaCache-get"></a>}}
\if{latex}{\out{\hypertarget{method-ZinaCache-get}{}}}
\subsection{Method \code{get()}}{
Retrieve data from cache
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZinaCache$get(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Cache key}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Cached data or NULL if not found
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZinaCache-has"></a>}}
\if{latex}{\out{\hypertarget{method-ZinaCache-has}{}}}
\subsection{Method \code{has()}}{
Check if key exists in cache
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZinaCache$has(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Cache key}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TRUE if exists
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZinaCache-remove"></a>}}
\if{latex}{\out{\hypertarget{method-ZinaCache-remove}{}}}
\subsection{Method \code{remove()}}{
Remove item from cache
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZinaCache$remove(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Cache key}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZinaCache-clear"></a>}}
\if{latex}{\out{\hypertarget{method-ZinaCache-clear}{}}}
\subsection{Method \code{clear()}}{
Clear all cache
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZinaCache$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZinaCache-stats"></a>}}
\if{latex}{\out{\hypertarget{method-ZinaCache-stats}{}}}
\subsection{Method \code{stats()}}{
Get cache statistics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZinaCache$stats()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List with cache stats
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZinaCache-keys"></a>}}
\if{latex}{\out{\hypertarget{method-ZinaCache-keys}{}}}
\subsection{Method \code{keys()}}{
List all cached keys
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZinaCache$keys()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character vector of keys
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZinaCache-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ZinaCache-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZinaCache$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
