% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{ZinaSuite-package}
\alias{ZinaSuite}
\alias{ZinaSuite-package}
\title{ZinaSuite: A Modern R Package for UCSC Xena Data Analysis}
\description{
Provides a modernized framework for interactively exploring
UCSC Xena datasets. It enables downloading, analyzing and visualizing
datasets from UCSC Xena with enhanced performance through mirai-based
asynchronous computing and modular Shiny application design.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/WangLabCSU/ZinaSuite}
  \item \url{https://wanglabcsu.github.io/ZinaSuite/}
  \item Report bugs at \url{https://github.com/WangLabCSU/ZinaSuite/issues}
}

}
\author{
\strong{Maintainer}: Shixiang Wang \email{w_shixiang@163.com} (\href{https://orcid.org/0000-0001-9855-7357}{ORCID})

}
\keyword{internal}
