% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pcawg.R
\name{PCAWGData}
\alias{PCAWGData}
\title{PCAWGData R6 Class}
\description{
R6 class for accessing PCAWG (Pan-Cancer Analysis of Whole Genomes) data.
}
\examples{
\dontrun{
# Create PCAWG data instance
pcawg <- PCAWGData$new()

# Query gene expression
tp53_expr <- pcawg$get_gene_expression("TP53")

# Query mutation status
tp53_mut <- pcawg$get_mutation_status("TP53")
}
}
\section{Super class}{
\code{\link[ZinaSuite:DataSource]{ZinaSuite::DataSource}} -> \code{PCAWGData}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PCAWGData-new}{\code{PCAWGData$new()}}
\item \href{#method-PCAWGData-get_gene_expression}{\code{PCAWGData$get_gene_expression()}}
\item \href{#method-PCAWGData-get_fusion}{\code{PCAWGData$get_fusion()}}
\item \href{#method-PCAWGData-get_mirna}{\code{PCAWGData$get_mirna()}}
\item \href{#method-PCAWGData-get_promoter}{\code{PCAWGData$get_promoter()}}
\item \href{#method-PCAWGData-get_apobec}{\code{PCAWGData$get_apobec()}}
\item \href{#method-PCAWGData-query}{\code{PCAWGData$query()}}
\item \href{#method-PCAWGData-clone}{\code{PCAWGData$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ZinaSuite" data-topic="DataSource" data-id="clear_cache"><a href='../../ZinaSuite/html/DataSource.html#method-DataSource-clear_cache'><code>ZinaSuite::DataSource$clear_cache()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ZinaSuite" data-topic="DataSource" data-id="get_info"><a href='../../ZinaSuite/html/DataSource.html#method-DataSource-get_info'><code>ZinaSuite::DataSource$get_info()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ZinaSuite" data-topic="DataSource" data-id="query_batch"><a href='../../ZinaSuite/html/DataSource.html#method-DataSource-query_batch'><code>ZinaSuite::DataSource$query_batch()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PCAWGData-new"></a>}}
\if{latex}{\out{\hypertarget{method-PCAWGData-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new PCAWGData instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PCAWGData$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PCAWGData-get_gene_expression"></a>}}
\if{latex}{\out{\hypertarget{method-PCAWGData-get_gene_expression}{}}}
\subsection{Method \code{get_gene_expression()}}{
Get gene expression data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PCAWGData$get_gene_expression(gene)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gene}}{Gene symbol}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Named numeric vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PCAWGData-get_fusion"></a>}}
\if{latex}{\out{\hypertarget{method-PCAWGData-get_fusion}{}}}
\subsection{Method \code{get_fusion()}}{
Get fusion data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PCAWGData$get_fusion(gene)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gene}}{Gene symbol}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Named numeric vector (0/1)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PCAWGData-get_mirna"></a>}}
\if{latex}{\out{\hypertarget{method-PCAWGData-get_mirna}{}}}
\subsection{Method \code{get_mirna()}}{
Get miRNA expression data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PCAWGData$get_mirna(mirna, norm = c("TMM", "UQ"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mirna}}{miRNA ID (e.g., "hsa-miR-21-5p")}

\item{\code{norm}}{Normalization method: "TMM" (default) or "UQ"}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Named numeric vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PCAWGData-get_promoter"></a>}}
\if{latex}{\out{\hypertarget{method-PCAWGData-get_promoter}{}}}
\subsection{Method \code{get_promoter()}}{
Get promoter activity data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PCAWGData$get_promoter(promoter, type = c("raw", "relative", "outlier"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{promoter}}{Promoter identifier (e.g., "prmtr.10000") or gene symbol}

\item{\code{type}}{Promoter activity type: "raw" (default), "relative", or "outlier"}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Named numeric vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PCAWGData-get_apobec"></a>}}
\if{latex}{\out{\hypertarget{method-PCAWGData-get_apobec}{}}}
\subsection{Method \code{get_apobec()}}{
Get APOBEC mutagenesis data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PCAWGData$get_apobec(
  identifier = c("tCa_MutLoad_MinEstimate", "APOBECtCa_enrich", "A3A_or_A3B",
    "APOBEC_tCa_enrich_quartile")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{identifier}}{APOBEC identifier}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Named numeric vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PCAWGData-query"></a>}}
\if{latex}{\out{\hypertarget{method-PCAWGData-query}{}}}
\subsection{Method \code{query()}}{
Query data by type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PCAWGData$query(identifier, data_type = "expression", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{identifier}}{Gene symbol or other identifier}

\item{\code{data_type}}{Type of data: "expression", "fusion", "miRNA", "promoter", "APOBEC"}

\item{\code{...}}{Additional parameters (e.g., norm for miRNA)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Named numeric vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PCAWGData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PCAWGData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PCAWGData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
