% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-advanced-logic.R, R/utils.R
\name{tcga_barcode_to_cancer}
\alias{tcga_barcode_to_cancer}
\title{TCGA Barcode to Cancer}
\usage{
tcga_barcode_to_cancer(barcode)

tcga_barcode_to_cancer(barcode)
}
\arguments{
\item{barcode}{TCGA barcode}
}
\value{
Cancer type

Cancer type abbreviation
}
\description{
TCGA Barcode to Cancer

Extract cancer type from TCGA barcode
}
\keyword{internal}
