% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-advanced-logic.R, R/utils.R
\name{tcga_barcode_to_type}
\alias{tcga_barcode_to_type}
\title{TCGA Barcode to Type}
\usage{
tcga_barcode_to_type(barcode)

tcga_barcode_to_type(barcode)
}
\arguments{
\item{barcode}{TCGA barcode}
}
\value{
Sample type

Sample type: "tumor" or "normal"
}
\description{
TCGA Barcode to Type

Convert TCGA barcode to sample type
}
\keyword{internal}
